Split related commands into their own files.
