+++ commands.go
@@ -1,24 +1,9 @@
 package main
 
 import (
-	"bytes"
-	"fmt"
-	"image"
-	"image/png"
-	"math/rand"
-	"net/http"
-	"net/url"
-	"regexp"
-	"sort"
-	"strconv"
 	"strings"
 
-	"github.com/JoshuaDoes/go-cve"
-	"github.com/JoshuaDoes/urbandictionary"
-	//prettyTime "github.com/andanhm/go-prettytime"
 	"github.com/bwmarrin/discordgo"
-	"github.com/disintegration/gift"
-	"github.com/rylio/ytdl"
 )
 
 // Command holds data related to a command executable by any message system
@@ -352,986 +337,4 @@ func getCommandUsage(commandName, title string) *discordgo.MessageEmbed {
 	usageEmbed.Fields = parameterFields
 
 	return usageEmbed
-}
-
-func commandHelp(args []string, env *CommandEnvironment) *discordgo.MessageEmbed {
-	//First see if help text is being requested for a particular command
-	if len(args) > 0 {
-		if command, exists := botData.Commands[args[0]]; exists {
-			if command.IsAlternateOf != "" {
-				if commandAlternate, exists := botData.Commands[command.IsAlternateOf]; exists {
-					command = commandAlternate
-				} else {
-					return nil
-				}
-			}
-			return getCommandUsage(args[0], "Help for **"+args[0]+"**")
-		}
-	}
-
-	//Before we fetch help text data, we need to have an alphabetical listing of commands
-	var commandMapKeys []string
-	for commandMapKey := range botData.Commands {
-		commandMapKeys = append(commandMapKeys, commandMapKey)
-	}
-	sort.Strings(commandMapKeys)
-
-	//Create a dynamic list of fields for the help embed
-	commandFields := []*discordgo.MessageEmbedField{}
-
-	//Iterate over the alphabetically sorted command list and add each listed command to the help embed field list
-	for _, commandName := range commandMapKeys {
-		command := botData.Commands[commandName]
-		if command.IsAlternateOf == "" {
-			if permissionsAllowed, _ := MemberHasPermission(botData.DiscordSession, env.Guild.ID, env.User.ID, discordgo.PermissionAdministrator|command.RequiredPermissions); permissionsAllowed || command.RequiredPermissions == 0 {
-				commandField := &discordgo.MessageEmbedField{Name: botData.CommandPrefix + commandName, Value: command.HelpText, Inline: true}
-				commandFields = append(commandFields, commandField)
-			}
-		}
-	}
-
-	//Create the help embed and give it the command list
-	helpEmbed := NewEmbed().
-		SetTitle(botData.BotName + " - Help").
-		SetDescription("A list of commands you have permission to use.").
-		SetColor(0xFAFAFA).MessageEmbed
-	helpEmbed.Fields = commandFields
-
-	//Return the help embed to the caller
-	return helpEmbed
-}
-func commandAbout(args []string, env *CommandEnvironment) *discordgo.MessageEmbed {
-	return NewEmbed().
-		SetTitle(botData.BotName+" - About").
-		SetDescription(botData.BotName+" is a Discord bot written in Google's Go programming language, intended for conversation and fact-based queries.").
-		AddField("How can I use "+botData.BotName+" in my server?", "Simply open the Invite Link at the end of this message and follow the on-screen instructions.").
-		AddField("How can I help keep "+botData.BotName+" running?", "The best ways to help keep "+botData.BotName+" running are to either donate using the Donation Link or contribute to the source code using the Source Code Link, both at the end of this message.").
-		AddField("How can I use "+botData.BotName+"?", "There are many ways to make use of "+botData.BotName+".\n1) Type ``cli$help`` and try using some of the available commands.\n2) Ask "+botData.BotName+" a question, ex: ``@"+botData.BotName+"#1823, what time is it?`` or ``@"+botData.BotName+"#1823, what is DiscordApp?``.").
-		AddField("Where can I join the "+botData.BotName+" Discord server?", "If you would like to get help and support with "+botData.BotName+" or experiment with the latest and greatest of "+botData.BotName+", use the Discord Server Invite Link at the end of this message.").
-		AddField("Invite Link", "https://discordapp.com/api/oauth2/authorize?client_id=374546169755598849&permissions=8&scope=bot").
-		AddField("Donation Link", "https://www.paypal.me/JoshuaDoes").
-		AddField("Source Code Link", "https://github.com/JoshuaDoes/clinet-discord/").
-		AddField("Discord Server Invite Link", "https://discord.gg/qkbKEWT").
-		SetColor(0x1C1C1C).MessageEmbed
-}
-func commandVersion(args []string, env *CommandEnvironment) *discordgo.MessageEmbed {
-	return NewEmbed().
-		SetTitle(botData.BotName+" - Version").
-		AddField("Build ID", BuildID).
-		AddField("Build Date", BuildDate).
-		AddField("Latest Development", GitCommitMsg).
-		AddField("GitHub Commit URL", GitHubCommitURL).
-		AddField("Golang Version", GolangVersion).
-		SetColor(0x1C1C1C).MessageEmbed
-}
-func commandCredits(args []string, env *CommandEnvironment) *discordgo.MessageEmbed {
-	return NewEmbed().
-		SetTitle(botData.BotName+" - Credits").
-		AddField("Bot Development", "- JoshuaDoes (2018)").
-		AddField("Programming Language", "- Golang").
-		AddField("Golang Libraries", "- https://github.com/bwmarrin/discordgo\n"+
-			"- https://github.com/disintegration/gift\n"+
-			"- https://github.com/JoshuaDoes/duckduckgolang\n"+
-			"- https://github.com/google/go-github/github\n"+
-			"- https://github.com/jonas747/dca\n"+
-			"- https://github.com/JoshuaDoes/go-soundcloud\n"+
-			"- https://github.com/JoshuaDoes/go-wolfram\n"+
-			"- https://github.com/koffeinsource/go-imgur\n"+
-			"- https://github.com/koffeinsource/go-klogger\n"+
-			"- https://github.com/nishanths/go-xkcd\n"+
-			"- https://github.com/paked/configure\n"+
-			"- https://github.com/robfig/cron\n"+
-			"- https://github.com/rylio/ytdl\n"+
-			"- https://google.golang.org/api/googleapi/transport\n"+
-			"- https://google.golang.org/api/youtube/v3").
-		AddField("Icon Design", "- thejsa").
-		AddField("Source Code", "- https://github.com/JoshuaDoes/clinet-discord").
-		SetColor(0x1C1C1C).MessageEmbed
-}
-func commandRoll(args []string, env *CommandEnvironment) *discordgo.MessageEmbed {
-	random := rand.Intn(6) + 1
-	return NewGenericEmbed("Roll", "You rolled a "+strconv.Itoa(random)+"!")
-}
-func commandDoubleRoll(args []string, env *CommandEnvironment) *discordgo.MessageEmbed {
-	random1 := rand.Intn(6) + 1
-	random2 := rand.Intn(6) + 1
-	randomTotal := random1 + random2
-	return NewGenericEmbed("Double Roll", "You rolled a "+strconv.Itoa(random1)+" and a "+strconv.Itoa(random2)+". The total is "+strconv.Itoa(randomTotal)+"!")
-}
-func commandCoinFlip(args []string, env *CommandEnvironment) *discordgo.MessageEmbed {
-	random := rand.Intn(2)
-	switch random {
-	case 0:
-		return NewGenericEmbed("Coin Flip", "The coin landed on heads!")
-	case 1:
-		return NewGenericEmbed("Coin Flip", "The coin landed on tails!")
-	}
-	return nil
-}
-func commandImage(args []string, env *CommandEnvironment) *discordgo.MessageEmbed {
-	if len(env.Message.Attachments) > 0 {
-		for _, attachment := range env.Message.Attachments {
-			srcImageURL := attachment.URL
-			srcImageHTTP, err := http.Get(srcImageURL)
-			if err != nil {
-				return NewErrorEmbed("Image Error", "Unable to fetch image.")
-			}
-			srcImage, _, err := image.Decode(srcImageHTTP.Body)
-			if err != nil {
-				return NewErrorEmbed("Image Error", "Unable to decode image.")
-			}
-
-			g := &gift.GIFT{}
-			var outImage bytes.Buffer
-
-			switch args[0] {
-			case "fliphorizontal":
-				g = gift.New(gift.FlipHorizontal())
-			case "flipvertical":
-				g = gift.New(gift.FlipVertical())
-			case "grayscale", "greyscale":
-				g = gift.New(gift.Grayscale())
-			case "invert":
-				g = gift.New(gift.Invert())
-			case "rotate90":
-				g = gift.New(gift.Rotate90())
-			case "rotate180":
-				g = gift.New(gift.Rotate180())
-			case "rotate270":
-				g = gift.New(gift.Rotate270())
-			case "sobel":
-				g = gift.New(gift.Sobel())
-			case "transpose":
-				g = gift.New(gift.Transpose())
-			case "transverse":
-				g = gift.New(gift.Transverse())
-			case "test":
-				g = gift.New(gift.Convolution(
-					[]float32{
-						-1, -1, 0,
-						-1, 1, 1,
-						0, 1, 1,
-					},
-					false, false, false, 0.0,
-				))
-			}
-
-			dstImage := image.NewRGBA(g.Bounds(srcImage.Bounds()))
-			g.Draw(dstImage, srcImage)
-
-			err = png.Encode(&outImage, dstImage)
-			if err != nil {
-				return NewErrorEmbed("Image Error", "Unable to encode processed image.")
-			}
-			_, err = botData.DiscordSession.ChannelMessageSendComplex(env.Channel.ID, &discordgo.MessageSend{
-				Content: "Processed image:",
-				File: &discordgo.File{
-					Name:   args[0] + ".png",
-					Reader: &outImage,
-				},
-				Embed: &discordgo.MessageEmbed{
-					Image: &discordgo.MessageEmbedImage{
-						URL: "attachment://" + args[0] + ".png",
-					},
-				},
-			})
-			if err != nil {
-				return NewErrorEmbed("Image Error", "Unable to upload processed image.")
-			}
-		}
-	} else {
-		return NewErrorEmbed("Image Error", "You must upload an image to process.")
-	}
-	return nil
-}
-
-func commandCVE(args []string, env *CommandEnvironment) *discordgo.MessageEmbed {
-	cveData, err := cve.GetCVE(args[0])
-	if err != nil {
-		return NewErrorEmbed("CVE Error", "There was an error fetching information about CVE ``"+args[0]+"``.")
-	}
-	return NewEmbed().
-		SetTitle(args[0]).
-		SetDescription(cveData.Summary).
-		SetColor(0xC93130).MessageEmbed
-}
-
-func commandXKCD(args []string, env *CommandEnvironment) *discordgo.MessageEmbed {
-	switch args[0] {
-	case "latest":
-		comic, err := botData.BotClients.XKCD.Latest()
-		if err != nil {
-			return NewErrorEmbed("XKCD Error", "There was an error fetching the latest XKCD comic.")
-		}
-		return NewEmbed().
-			SetTitle("xkcd - #" + strconv.Itoa(comic.Number)).
-			SetDescription(comic.Title).
-			SetImage(comic.ImageURL).
-			SetColor(0x96A8C8).MessageEmbed
-	case "random":
-		comic, err := botData.BotClients.XKCD.Random()
-		if err != nil {
-			return NewErrorEmbed("XKCD Error", "There was an error fetching a random XKCD comic.")
-		}
-		return NewEmbed().
-			SetTitle("xkcd - #" + strconv.Itoa(comic.Number)).
-			SetDescription(comic.Title).
-			SetImage(comic.ImageURL).
-			SetColor(0x96A8C8).MessageEmbed
-	default:
-		comicNumber, err := strconv.Atoi(args[0])
-		if err != nil {
-			return NewErrorEmbed("XKCD Error", "``"+args[0]+"`` is not a valid number.")
-		}
-
-		comic, err := botData.BotClients.XKCD.Get(comicNumber)
-		if err != nil {
-			return NewErrorEmbed("XKCD Error", "There was an error fetching XKCD comic #"+args[0]+".")
-		}
-		return NewEmbed().
-			SetTitle("xkcd - #" + args[0]).
-			SetDescription(comic.Title).
-			SetImage(comic.ImageURL).
-			SetColor(0x96A8C8).MessageEmbed
-	}
-}
-
-func commandImgur(args []string, env *CommandEnvironment) *discordgo.MessageEmbed {
-	responseEmbed, err := queryImgur(args[0])
-	if err != nil {
-		return NewErrorEmbed("Imgur Error", "There was an error fetching information about the specified URL.")
-	}
-	return responseEmbed
-}
-
-func commandGitHub(args []string, env *CommandEnvironment) *discordgo.MessageEmbed {
-	request := strings.Split(args[0], "/")
-
-	switch len(request) {
-	case 1: //Only user was specified
-		user, err := GitHubFetchUser(request[0])
-		if err != nil {
-			return NewErrorEmbed("GitHub Error", "There was an error fetching information about the specified user.")
-		}
-
-		fields := []*discordgo.MessageEmbedField{}
-
-		//Gather user info
-		if user.Bio != nil {
-			fields = append(fields, &discordgo.MessageEmbedField{Name: "Bio", Value: *user.Bio})
-		}
-		fields = append(fields, &discordgo.MessageEmbedField{Name: "Username", Value: *user.Login})
-		if user.Name != nil {
-			fields = append(fields, &discordgo.MessageEmbedField{Name: "Name", Value: *user.Name})
-		}
-		if user.Company != nil {
-			fields = append(fields, &discordgo.MessageEmbedField{Name: "Company", Value: *user.Company})
-		}
-		if *user.Blog != "" {
-			fields = append(fields, &discordgo.MessageEmbedField{Name: "Blog", Value: *user.Blog})
-		}
-		if user.Location != nil {
-			fields = append(fields, &discordgo.MessageEmbedField{Name: "Location", Value: *user.Location})
-		}
-		fields = append(fields, &discordgo.MessageEmbedField{Name: "Public Repos", Value: strconv.Itoa(*user.PublicRepos)})
-		fields = append(fields, &discordgo.MessageEmbedField{Name: "Public Gists", Value: strconv.Itoa(*user.PublicGists)})
-		fields = append(fields, &discordgo.MessageEmbedField{Name: "Following", Value: strconv.Itoa(*user.Following)})
-		fields = append(fields, &discordgo.MessageEmbedField{Name: "Followers", Value: strconv.Itoa(*user.Followers)})
-		fields = append(fields, &discordgo.MessageEmbedField{Name: "GitHub URL", Value: *user.HTMLURL})
-
-		//Build embed about user
-		responseEmbed := NewEmbed().
-			SetTitle("GitHub User: " + *user.Login).
-			SetImage(*user.AvatarURL).
-			SetColor(0x24292D).MessageEmbed
-		responseEmbed.Fields = fields
-
-		return responseEmbed
-	case 2: //Repo was specified
-		repo, err := GitHubFetchRepo(request[0], request[1])
-		if err != nil {
-			return NewErrorEmbed("GitHub Error", "There was an error fetchign information about the specified repo.")
-		}
-
-		fields := []*discordgo.MessageEmbedField{}
-
-		//Gather repo info
-		if repo.Description != nil && *repo.Description != "" {
-			fields = append(fields, &discordgo.MessageEmbedField{Name: "Description", Value: *repo.Description})
-		}
-		fields = append(fields, &discordgo.MessageEmbedField{Name: "Name", Value: *repo.FullName})
-		if repo.Homepage != nil && *repo.Homepage != "" {
-			fields = append(fields, &discordgo.MessageEmbedField{Name: "Homepage", Value: *repo.Homepage})
-		}
-		if len(repo.Topics) > 0 {
-			fields = append(fields, &discordgo.MessageEmbedField{Name: "Topics", Value: strings.Join(repo.Topics, ", ")})
-		}
-		fields = append(fields, &discordgo.MessageEmbedField{Name: "Default Branch", Value: *repo.DefaultBranch})
-		fields = append(fields, &discordgo.MessageEmbedField{Name: "Is Fork", Value: strconv.FormatBool(*repo.Fork)})
-		fields = append(fields, &discordgo.MessageEmbedField{Name: "Forks", Value: strconv.Itoa(*repo.ForksCount)})
-		fields = append(fields, &discordgo.MessageEmbedField{Name: "Networks", Value: strconv.Itoa(*repo.NetworkCount)})
-		fields = append(fields, &discordgo.MessageEmbedField{Name: "Open Issues", Value: strconv.Itoa(*repo.OpenIssuesCount)})
-		fields = append(fields, &discordgo.MessageEmbedField{Name: "Stargazers", Value: strconv.Itoa(*repo.StargazersCount)})
-		fields = append(fields, &discordgo.MessageEmbedField{Name: "Subscribers", Value: strconv.Itoa(*repo.SubscribersCount)})
-		fields = append(fields, &discordgo.MessageEmbedField{Name: "Watchers", Value: strconv.Itoa(*repo.WatchersCount)})
-		fields = append(fields, &discordgo.MessageEmbedField{Name: "GitHub URL", Value: *repo.HTMLURL})
-		fields = append(fields, &discordgo.MessageEmbedField{Name: "Clone URL", Value: *repo.CloneURL})
-		fields = append(fields, &discordgo.MessageEmbedField{Name: "Git URL", Value: *repo.GitURL})
-
-		//Build embed about repo
-		responseEmbed := NewEmbed().
-			SetTitle("GitHub Repo: " + *repo.FullName).
-			SetColor(0x24292D).MessageEmbed
-		responseEmbed.Fields = fields
-
-		return responseEmbed
-	}
-
-	return NewErrorEmbed("GitHub Error", "You must specify a GitHub user or a GitHub repo to fetch info about.\n\nExamples:\n```"+botData.CommandPrefix+"github JoshuaDoes\n"+botData.CommandPrefix+"gh JoshuaDoes/clinet-discord```")
-}
-
-func commandUrbanDictionary(args []string, env *CommandEnvironment) *discordgo.MessageEmbed {
-	results, err := urbandictionary.Query(args[0])
-	if err != nil {
-		return NewErrorEmbed("Urban Dictionary Error", "There was an error getting a result for that term.")
-	}
-
-	linkExp := regexp.MustCompile(`\[([^\]]*)\]`)
-	linkExpFunc := func(s string) string {
-		ss := linkExp.FindStringSubmatch(s)
-		if len(ss) == 0 {
-			return s
-		}
-		hyperlink := "https://www.urbandictionary.com/define.php?term=" + url.QueryEscape(ss[1])
-		return fmt.Sprintf("%s(%s)", s, hyperlink)
-	}
-
-	result := results.Results[0]
-	resultEmbed := NewEmbed().
-		SetTitle("Urban Dictionary - "+result.Word).
-		SetDescription(linkExp.ReplaceAllStringFunc(result.Definition, linkExpFunc)).
-		AddField("Example", linkExp.ReplaceAllStringFunc(result.Example, linkExpFunc)).
-		AddField("Author", result.Author).
-		AddField("Stats", "\U0001f44d "+strconv.Itoa(result.ThumbsUp)+" \U0001f44e "+strconv.Itoa(result.ThumbsDown)).
-		SetFooter("Results from Urban Dictionary.", "https://res.cloudinary.com/hrscywv4p/image/upload/c_limit,fl_lossy,h_300,w_300,f_auto,q_auto/v1/1194347/vo5ge6mdw4creyrgaq2m.png").
-		SetURL(result.Permalink).
-		SetColor(0xDC2A26).MessageEmbed
-
-	//Oddly enough, this isn't wanting to return anything for the moment so it's staying commented
-	//	date := prettyTime.Format(result.Date)
-	//	if date != "" {
-	//		resultEmbed.Fields = append(resultEmbed.Fields, &discordgo.MessageEmbedField{Name: "Date Written", Value: date})
-	//	}
-
-	return resultEmbed
-}
-
-func commandVoiceJoin(args []string, env *CommandEnvironment) *discordgo.MessageEmbed {
-	for _, voiceState := range env.Guild.VoiceStates {
-		if voiceState.UserID == env.Message.Author.ID {
-			voiceJoin(botData.DiscordSession, env.Guild.ID, voiceState.ChannelID, env.Message.ID)
-			return NewGenericEmbed("Voice", "Joined the voice channel.")
-		}
-	}
-	return NewErrorEmbed("Voice Error", "You must join the voice channel to use before using the join command.")
-}
-
-func commandVoiceLeave(args []string, env *CommandEnvironment) *discordgo.MessageEmbed {
-	for _, voiceState := range env.Guild.VoiceStates {
-		if voiceState.UserID == env.Message.Author.ID {
-			if voiceIsStreaming(env.Guild.ID) {
-				voiceStop(env.Guild.ID)
-			}
-			err := voiceLeave(env.Guild.ID, voiceState.ChannelID)
-			if err != nil {
-				return NewErrorEmbed("Voice Error", "There was an error leaving the voice channel.")
-			}
-			return NewGenericEmbed("Voice", "Left the voice channel.")
-		}
-	}
-	return NewErrorEmbed("Voice Error", "You must join the voice channel to use before using the leave command.")
-}
-
-func commandPlay(args []string, env *CommandEnvironment) *discordgo.MessageEmbed {
-	if env.UpdatedMessageEvent {
-		//Todo: Remove this once I figure out how to detect if message update was user-triggered
-		//Reason: If you use a YouTube/SoundCloud URL, Discord automatically updates the message with an embed
-		//As far as I know, bots have no way to know if this was a Discord- or user-triggered message update
-		//I eventually want users to be able to edit their play command to change a now playing or a queue entry that was misspelled
-		return nil
-	}
-
-	for guildData[env.Guild.ID].VoiceData.IsPlaybackPreparing {
-		//Wait for the handling of a previous playback command to finish
-	}
-	foundVoiceChannel := false
-	for _, voiceState := range env.Guild.VoiceStates {
-		if voiceState.UserID == env.Message.Author.ID {
-			foundVoiceChannel = true
-			voiceJoin(botData.DiscordSession, env.Guild.ID, voiceState.ChannelID, env.Message.ID)
-			break
-		}
-	}
-	if !foundVoiceChannel {
-		return NewErrorEmbed("Voice Error", "You must join the voice channel to use before using the play command.")
-	}
-	//Prevent other play commands in this voice session from messing up this process
-	guildData[env.Guild.ID].VoiceData.IsPlaybackPreparing = true
-
-	if len(args) >= 1 { //Query or URL was specified
-		_, err := url.ParseRequestURI(args[0]) //Check to see if the first parameter is a URL
-		if err != nil {                        //First parameter is not a URL
-			queryURL, err := voiceGetQuery(strings.Join(args, " "))
-			if err != nil {
-				guildData[env.Guild.ID].VoiceData.IsPlaybackPreparing = false //We're done so we should allow the next play command to run
-				return NewErrorEmbed("Voice Error", "There was an error getting a result for the specified query.")
-			}
-			queueData := AudioQueueEntry{MediaURL: queryURL, Requester: env.Message.Author, Type: "youtube"}
-			queueData.FillMetadata()
-			if voiceIsStreaming(env.Guild.ID) {
-				guildData[env.Guild.ID].QueueAdd(queueData)
-				guildData[env.Guild.ID].VoiceData.IsPlaybackPreparing = false //We're done so we should allow the next play command to run
-				return queueData.GetQueueAddedEmbed()
-			}
-			guildData[env.Guild.ID].AudioNowPlaying = queueData
-			guildData[env.Guild.ID].VoiceData.IsPlaybackPreparing = false //We're done so we should allow the next play command to run
-			go voicePlayWrapper(botData.DiscordSession, env.Guild.ID, env.Channel.ID, queueData.MediaURL)
-			return queueData.GetNowPlayingEmbed()
-		}
-
-		//First parameter is a URL
-		queueData := AudioQueueEntry{MediaURL: args[0], Requester: env.Message.Author}
-		queueData.FillMetadata()
-		if voiceIsStreaming(env.Guild.ID) {
-			guildData[env.Guild.ID].QueueAdd(queueData)
-			guildData[env.Guild.ID].VoiceData.IsPlaybackPreparing = false //We're done so we should allow the next play command to run
-			return queueData.GetQueueAddedEmbed()
-		}
-		guildData[env.Guild.ID].AudioNowPlaying = queueData
-		guildData[env.Guild.ID].VoiceData.IsPlaybackPreparing = false //We're done so we should allow the next play command to run
-		go voicePlayWrapper(botData.DiscordSession, env.Guild.ID, env.Channel.ID, queueData.MediaURL)
-		return queueData.GetNowPlayingEmbed()
-	}
-
-	if voiceIsStreaming(env.Guild.ID) {
-		if len(env.Message.Attachments) > 0 {
-			for _, attachment := range env.Message.Attachments {
-				queueData := AudioQueueEntry{MediaURL: attachment.URL, Requester: env.Message.Author}
-				queueData.FillMetadata()
-				guildData[env.Guild.ID].QueueAdd(queueData)
-			}
-			guildData[env.Guild.ID].VoiceData.IsPlaybackPreparing = false //We're done so we should allow the next play command to run
-			return NewGenericEmbed("Voice", "Added the attached files to the guild queue.")
-		}
-		isPaused, _ := voiceGetPauseState(env.Guild.ID)
-		if isPaused {
-			_, _ = voiceResume(env.Guild.ID)
-			guildData[env.Guild.ID].VoiceData.IsPlaybackPreparing = false //We're done so we should allow the next play command to run
-			return NewGenericEmbed("Voice", "Resumed the audio playback.")
-		}
-		guildData[env.Guild.ID].VoiceData.IsPlaybackPreparing = false //We're done so we should allow the next play command to run
-		return NewErrorEmbed("Voice Error", "Already playing audio.")
-	}
-	if len(env.Message.Attachments) > 0 {
-		for _, attachment := range env.Message.Attachments {
-			queueData := AudioQueueEntry{MediaURL: attachment.URL, Requester: env.Message.Author}
-			queueData.FillMetadata()
-			guildData[env.Guild.ID].QueueAdd(queueData)
-		}
-		return NewGenericEmbed("Voice", "Added the attached files to the guild queue. Use ``"+botData.CommandPrefix+"play`` to begin playback from the beginning of the queue.")
-	}
-	if guildData[env.Guild.ID].AudioNowPlaying.MediaURL != "" {
-		queueData := guildData[env.Guild.ID].AudioNowPlaying
-		queueData.FillMetadata()
-		guildData[env.Guild.ID].VoiceData.IsPlaybackPreparing = false //We're done so we should allow the next play command to run
-		go voicePlayWrapper(botData.DiscordSession, env.Guild.ID, env.Message.ChannelID, queueData.MediaURL)
-		return queueData.GetQueueAddedEmbed()
-	}
-	if len(guildData[env.Guild.ID].AudioQueue) > 0 {
-		queueData := guildData[env.Guild.ID].AudioQueue[0]
-		queueData.FillMetadata()
-		guildData[env.Guild.ID].QueueRemove(0)
-		go voicePlayWrapper(botData.DiscordSession, env.Guild.ID, env.Message.ChannelID, queueData.MediaURL)
-		guildData[env.Guild.ID].VoiceData.IsPlaybackPreparing = false //We're done so we should allow the next play command to run
-		return queueData.GetQueueAddedEmbed()
-	}
-
-	guildData[env.Guild.ID].VoiceData.IsPlaybackPreparing = false //We're done so we should allow the next play command to run
-	return NewGenericEmbed("Voice Error", "Some kind of strange logic flow occurred. Consider sending this to a developer.")
-}
-
-func commandStop(args []string, env *CommandEnvironment) *discordgo.MessageEmbed {
-	for _, voiceState := range env.Guild.VoiceStates {
-		if voiceState.UserID == env.Message.Author.ID {
-			if voiceIsStreaming(env.Guild.ID) {
-				voiceStop(env.Guild.ID)
-				return NewGenericEmbed("Voice", "Stopped the audio playback.")
-			}
-			return NewErrorEmbed("Voice Error", "There is no audio currently playing.")
-		}
-	}
-	return NewErrorEmbed("Voice Error", "You must join the voice channel "+botData.BotName+" to use before using the "+env.Command+" command.")
-}
-
-func commandSkip(args []string, env *CommandEnvironment) *discordgo.MessageEmbed {
-	for _, voiceState := range env.Guild.VoiceStates {
-		if voiceState.UserID == env.Message.Author.ID {
-			if voiceIsStreaming(env.Guild.ID) {
-				voiceSkip(env.Guild.ID)
-				return nil
-			}
-			return NewErrorEmbed("Voice Error", "There is no audio currently playing.")
-		}
-	}
-	return NewErrorEmbed("Voice Error", "You must join the voice channel "+botData.BotName+" to use before using the "+env.Command+" command.")
-}
-
-func commandPause(args []string, env *CommandEnvironment) *discordgo.MessageEmbed {
-	for _, voiceState := range env.Guild.VoiceStates {
-		if voiceState.UserID == env.Message.Author.ID {
-			isPaused, err := voicePause(env.Guild.ID)
-			if err != nil {
-				if isPaused {
-					return NewErrorEmbed("Voice Error", "Already playing audio.")
-				}
-				return NewErrorEmbed("Voice Error", "There is no audio currently playing.")
-			}
-			return NewGenericEmbed("Voice", "Paused the audio playback.")
-		}
-	}
-	return NewErrorEmbed("Voice Error", "You must join the voice channel "+botData.BotName+" to use before using the "+env.Command+" command.")
-}
-
-func commandResume(args []string, env *CommandEnvironment) *discordgo.MessageEmbed {
-	for _, voiceState := range env.Guild.VoiceStates {
-		if voiceState.UserID == env.Message.Author.ID {
-			isPaused, err := voiceResume(env.Guild.ID)
-			if err != nil {
-				if isPaused {
-					return NewErrorEmbed("Voice Error", "Already playing audio.")
-				}
-				return NewErrorEmbed("Voice Error", "There is no audio currently playing.")
-			}
-			return NewGenericEmbed("Voice", "Resumed the audio playback.")
-		}
-	}
-	return NewErrorEmbed("Voice Error", "You must join the voice channel "+botData.BotName+" to use before using the "+env.Command+" command.")
-}
-
-func commandVolume(args []string, env *CommandEnvironment) *discordgo.MessageEmbed {
-	volume, err := strconv.Atoi(args[0])
-	if err != nil {
-		return NewErrorEmbed("Volume Error", "``"+args[0]+"`` is not a valid number.")
-	}
-
-	if volume < 0 || volume > 512 {
-		return NewErrorEmbed("Volume Error", "You must specify a volume level from 0 to 512, with 256 being normal volume.")
-	}
-
-	if guildData[env.Guild.ID].VoiceData.EncodingOptions == nil {
-		guildData[env.Guild.ID].VoiceData.EncodingOptions = encodeOptionsPresetHigh
-	}
-	guildData[env.Guild.ID].VoiceData.EncodingOptions.Volume = volume
-	return NewErrorEmbed("Volume", "Set the volume for audio playback to "+args[0]+".")
-}
-
-func commandRepeat(args []string, env *CommandEnvironment) *discordgo.MessageEmbed {
-	switch guildData[env.Guild.ID].VoiceData.RepeatLevel {
-	case 0: //No repeat
-		guildData[env.Guild.ID].VoiceData.RepeatLevel = 1
-		return NewGenericEmbed("Voice", "The queue will be repeated on a loop.")
-	case 1: //Repeat the current queue
-		guildData[env.Guild.ID].VoiceData.RepeatLevel = 2
-		return NewGenericEmbed("Voice", "The now playing entry will be repeated on a loop.")
-	case 2: //Repeat what's in the now playing slot
-		guildData[env.Guild.ID].VoiceData.RepeatLevel = 0
-		return NewGenericEmbed("Voice", "The queue will play through as normal.")
-	}
-	return nil
-}
-
-func commandShuffle(args []string, env *CommandEnvironment) *discordgo.MessageEmbed {
-	guildData[env.Guild.ID].VoiceData.Shuffle = !guildData[env.Guild.ID].VoiceData.Shuffle
-	if guildData[env.Guild.ID].VoiceData.Shuffle {
-		return NewGenericEmbed("Voice", "The queue will be shuffled around in a random order while playing.")
-	}
-	return NewGenericEmbed("Voice", "The queue will play through as normal.")
-}
-
-func commandYouTube(args []string, env *CommandEnvironment) *discordgo.MessageEmbed {
-	for guildData[env.Guild.ID].VoiceData.IsPlaybackPreparing {
-		//Wait for the handling of a previous playback command to finish
-	}
-	foundVoiceChannel := false
-	for _, voiceState := range env.Guild.VoiceStates {
-		if voiceState.UserID == env.Message.Author.ID {
-			foundVoiceChannel = true
-			voiceJoin(botData.DiscordSession, env.Guild.ID, voiceState.ChannelID, env.Message.ID)
-			break
-		}
-	}
-	if !foundVoiceChannel {
-		return NewErrorEmbed("YouTube Error", "You must join the voice channel to use before using the "+env.Command+" command.")
-	}
-
-	page := &YouTubeResultNav{}
-
-	switch args[0] {
-	case "search", "s":
-		query := strings.Join(args[1:], " ")
-		if query == "" {
-			return NewErrorEmbed("YouTube Error", "You must enter a search query to use before using the "+args[0]+" command.")
-		}
-
-		if guildData[env.Guild.ID].YouTubeResults == nil {
-			guildData[env.Guild.ID].YouTubeResults = make(map[string]*YouTubeResultNav)
-		}
-
-		guildData[env.Guild.ID].YouTubeResults[env.Message.Author.ID] = &YouTubeResultNav{}
-
-		page = guildData[env.Guild.ID].YouTubeResults[env.Message.Author.ID]
-		err := page.Search(query)
-		if err != nil {
-			return NewErrorEmbed("YouTube Error", "There was an error getting a result for the specified query.")
-		}
-	case "next", "n", "forward", "+":
-		if guildData[env.Guild.ID].YouTubeResults == nil {
-			return NewErrorEmbed("YouTube Error", "No search session is in progress.")
-		}
-
-		page = guildData[env.Guild.ID].YouTubeResults[env.Message.Author.ID]
-		err := page.Next()
-		if err != nil {
-			return NewErrorEmbed("YouTube Error", "There was an error finding the next page.")
-		}
-	case "prev", "previous", "p", "back", "-":
-		if guildData[env.Guild.ID].YouTubeResults == nil {
-			return NewErrorEmbed("YouTube Error", "No search session is in progress.")
-		}
-
-		page = guildData[env.Guild.ID].YouTubeResults[env.Message.Author.ID]
-		err := page.Prev()
-		if err != nil {
-			return NewErrorEmbed("YouTube Error", "There was an error finding the previous page.")
-		}
-	case "cancel", "c":
-		if guildData[env.Guild.ID].YouTubeResults[env.Message.Author.ID] != nil {
-			guildData[env.Guild.ID].YouTubeResults[env.Message.Author.ID] = nil
-			return NewGenericEmbed("YouTube", "Cancelled the search session.")
-		}
-		return NewErrorEmbed("YouTube Error", "No search session is in progress.")
-	case "select", "choose":
-		if guildData[env.Guild.ID].YouTubeResults == nil {
-			return NewErrorEmbed("YouTube Error", "No search session is in progress.")
-		}
-		if len(args) < 2 {
-			return NewErrorEmbed("YouTube Error", "You must specify which search result to select.")
-		}
-
-		page = guildData[env.Guild.ID].YouTubeResults[env.Message.Author.ID]
-		results, _ := page.GetResults()
-
-		selection, err := strconv.Atoi(args[1])
-		if err != nil {
-			return NewErrorEmbed("YouTube Error", "``"+args[1]+"`` is not a valid number.")
-		}
-		if selection > len(results) || selection <= 0 {
-			return NewErrorEmbed("YouTube Error", "An invalid selection was specified.")
-		}
-
-		result := results[selection-1]
-		resultURL := "https://youtube.com/watch?v=" + result.Id.VideoId
-
-		queueData := AudioQueueEntry{MediaURL: resultURL, Requester: env.Message.Author, Type: "youtube"}
-		queueData.FillMetadata()
-		if voiceIsStreaming(env.Guild.ID) {
-			guildData[env.Guild.ID].QueueAdd(queueData)
-			return queueData.GetQueueAddedEmbed()
-		}
-		guildData[env.Guild.ID].AudioNowPlaying = queueData
-		go voicePlayWrapper(botData.DiscordSession, env.Guild.ID, env.Channel.ID, queueData.MediaURL)
-		return queueData.GetNowPlayingEmbed()
-	default:
-		return NewErrorEmbed("YouTube Error", "Unknown command ``"+args[0]+"``.")
-	}
-
-	commandList := botData.CommandPrefix + env.Command + " select N - Selects result N"
-	if page.PrevPageToken != "" {
-		commandList += "\n" + botData.CommandPrefix + env.Command + " prev - Displays the results for the previous page"
-	}
-	if page.NextPageToken != "" {
-		commandList += "\n" + botData.CommandPrefix + env.Command + " next - Displays the results for the next page"
-	}
-	commandListField := &discordgo.MessageEmbedField{Name: "Commands", Value: commandList}
-
-	results, _ := page.GetResults()
-	responseEmbed := NewEmbed().
-		SetTitle("YouTube Search Results - Page " + strconv.Itoa(page.PageNumber)).
-		SetDescription(strconv.FormatInt(page.TotalResults, 10) + " results for \"" + page.Query + "\"").
-		SetColor(0xFF0000).MessageEmbed
-
-	fields := []*discordgo.MessageEmbedField{}
-	for i := 0; i < len(results); i++ {
-		videoInfo, err := ytdl.GetVideoInfo("https://youtube.com/watch?v=" + results[i].Id.VideoId)
-		if err == nil {
-			author := videoInfo.Author
-			title := videoInfo.Title
-
-			fields = append(fields, &discordgo.MessageEmbedField{Name: "Result #" + strconv.Itoa(i+1), Value: "[" + title + "](https://youtube.com/watch?v=" + results[i].Id.VideoId + ") by **" + author + "**"})
-		}
-	}
-	fields = append(fields, commandListField)
-	responseEmbed.Fields = fields
-
-	return responseEmbed
-}
-
-func commandQueue(args []string, env *CommandEnvironment) *discordgo.MessageEmbed {
-	if len(args) >= 1 {
-		switch args[0] {
-		case "clear":
-			if len(guildData[env.Guild.ID].AudioQueue) > 0 {
-				guildData[env.Guild.ID].QueueClear()
-				return NewGenericEmbed("Queue", "Cleared the queue.")
-			}
-			return NewErrorEmbed("Queue Error", "There are no entries in the queue to clear.")
-		case "remove":
-			if len(args) == 1 {
-				return NewErrorEmbed("Queue Error", "You must specify which queue entries to remove.")
-			}
-
-			for _, queueEntry := range args[1:] {
-				queueEntryNumber, err := strconv.Atoi(queueEntry)
-				if err != nil {
-					return NewErrorEmbed("Queue Error", "``"+queueEntry+"`` is not a valid number.")
-				}
-				queueEntryNumber--
-
-				if queueEntryNumber > len(guildData[env.Guild.ID].AudioQueue) || queueEntryNumber < 0 {
-					return NewErrorEmbed("Queue Error", "``"+queueEntry+"`` is not a valid queue entry.")
-				}
-			}
-
-			var newAudioQueue []AudioQueueEntry
-			for queueEntryN, queueEntry := range guildData[env.Guild.ID].AudioQueue {
-				keepQueueEntry := true
-				for _, removedQueueEntry := range args[1:] {
-					removedQueueEntryNumber, _ := strconv.Atoi(removedQueueEntry)
-					removedQueueEntryNumber--
-					if queueEntryN == removedQueueEntryNumber {
-						keepQueueEntry = false
-						break
-					}
-				}
-				if keepQueueEntry {
-					newAudioQueue = append(newAudioQueue, queueEntry)
-				}
-			}
-
-			guildData[env.Guild.ID].AudioQueue = newAudioQueue
-
-			if len(args) > 2 {
-				return NewGenericEmbed("Queue", "Successfully removed the specified queue entries.")
-			}
-			return NewGenericEmbed("Queue", "Successfully removed the specified queue entry.")
-		}
-	}
-
-	if len(guildData[env.Guild.ID].AudioQueue) == 0 {
-		return NewErrorEmbed("Queue Error", "There are no entries in the queue.")
-	}
-	queueList := ""
-	for queueEntryNumber, queueEntry := range guildData[env.Guild.ID].AudioQueue {
-		displayNumber := strconv.Itoa(queueEntryNumber + 1)
-		if queueList != "" {
-			queueList += "\n"
-		}
-		switch queueEntry.Type {
-		case "youtube", "soundcloud":
-			queueList += displayNumber + ". [" + queueEntry.Title + "](" + queueEntry.MediaURL + ") by **" + queueEntry.Author + "** | Requested by <@" + queueEntry.Requester.ID + ">"
-		default:
-			queueList += displayNumber + ". " + queueEntry.MediaURL + " | Requested by <@" + queueEntry.Requester.ID + ">"
-		}
-	}
-	return NewGenericEmbed("Queue for "+env.Guild.Name, queueList)
-}
-
-func commandNowPlaying(args []string, env *CommandEnvironment) *discordgo.MessageEmbed {
-	if voiceIsStreaming(env.Guild.ID) {
-		return guildData[env.Guild.ID].AudioNowPlaying.GetNowPlayingDurationEmbed(guildData[env.Guild.ID].VoiceData.StreamingSession)
-	}
-	return NewErrorEmbed("Now Playing Error", "There is no audio currently playing.")
-}
-
-//Moderator commands
-func commandPurge(args []string, env *CommandEnvironment) *discordgo.MessageEmbed {
-	amount, err := strconv.Atoi(args[0])
-	if err != nil {
-		return NewErrorEmbed("Purge Error", "``"+args[0]+"`` is not a valid number.")
-	}
-	if amount <= 0 || amount > 100 {
-		return NewErrorEmbed("Purge Error", "Amount of messages to purge must be between 1 and 100.")
-	}
-
-	messages, err := botData.DiscordSession.ChannelMessages(env.Channel.ID, amount, env.Message.ID, "", "")
-	if err != nil {
-		return NewErrorEmbed("Purge Error", "An error occurred fetching the last "+args[0]+" messages.")
-	}
-
-	messageIDs := make([]string, 0)
-
-	if len(args) > 1 {
-		for i := 0; i < len(messages); i++ {
-			for j := 0; j < len(env.Message.Mentions); j++ {
-				if env.Message.Mentions[j].ID == messages[i].Author.ID {
-					messageIDs = append(messageIDs, messages[i].ID)
-					break
-				}
-			}
-		}
-
-		err = botData.DiscordSession.ChannelMessagesBulkDelete(env.Channel.ID, messageIDs)
-		if err != nil {
-			return NewErrorEmbed("Purge Error", "An error occurred deleting the last "+args[0]+" messages from the specified user(s).")
-		}
-
-		return NewGenericEmbed("Purge", "Successfully purged the last "+args[0]+" messages from the specified user(s).")
-	}
-
-	for i := 0; i < len(messages); i++ {
-		messageIDs = append(messageIDs, messages[i].ID)
-	}
-
-	err = botData.DiscordSession.ChannelMessagesBulkDelete(env.Channel.ID, messageIDs)
-	if err != nil {
-		return NewErrorEmbed("Purge Error", "An error occurred deleting the last "+args[0]+" messages.")
-	}
-
-	return NewGenericEmbed("Purge", "Successfully purged the last "+args[0]+" messages.")
-}
-func commandKick(args []string, env *CommandEnvironment) *discordgo.MessageEmbed {
-	if len(env.Message.Mentions) == 0 {
-		NewErrorEmbed("Kick Error", "You must specify which user(s) to kick from the server.")
-	}
-
-	reasonMessage := ""
-	usersToKick := make([]string, 0)
-	for i, part := range args {
-		if strings.HasPrefix(part, "<@") && strings.HasSuffix(part, ">") {
-			usersToKick = append(usersToKick, strings.TrimRight(strings.TrimLeft(part, "<@!"), ">"))
-			continue
-		}
-		reasonMessage = strings.Join(args[i:], " ")
-		break
-	}
-	if len(usersToKick) == 0 {
-		return NewErrorEmbed("Kick Error", "You must specify which user(s) to kick.")
-	}
-
-	if reasonMessage == "" {
-		for i := range usersToKick {
-			err := botData.DiscordSession.GuildMemberDelete(env.Guild.ID, usersToKick[i])
-			if err != nil {
-				return NewErrorEmbed("Kick Error", "An error occurred kicking <@"+usersToKick[i]+">. Please consider manually kicking and report this issue to a developer.")
-			}
-		}
-		return NewGenericEmbed("Kick", "Successfully kicked the selected user(s).")
-	}
-	for i := range usersToKick {
-		err := botData.DiscordSession.GuildMemberDeleteWithReason(env.Guild.ID, usersToKick[i], reasonMessage)
-		if err != nil {
-			return NewErrorEmbed("Kick Error", "An error occurred kicking <@"+usersToKick[i]+">. Please consider manually kicking and report this issue to a developer.")
-		}
-	}
-	return NewGenericEmbed("Kick", "Successfully kicked the selected user(s) for the following reason:\n**"+reasonMessage+"**")
-}
-func commandBan(args []string, env *CommandEnvironment) *discordgo.MessageEmbed {
-	if len(env.Message.Mentions) == 0 {
-		return NewErrorEmbed("Ban Error", "You must specify which user(s) to ban from the server.")
-	}
-
-	reasonMessage := ""
-	usersToBan := make([]string, 0)
-	messagesDaysToDelete := 0
-	for i, part := range args {
-		if i == 0 {
-			days, err := strconv.Atoi(part)
-			if err == nil {
-				messagesDaysToDelete = days
-				continue
-			} else {
-				return NewErrorEmbed("Ban Error", "``"+part+"`` is not a valid number.")
-			}
-		}
-		if strings.HasPrefix(part, "<@") && strings.HasSuffix(part, ">") {
-			usersToBan = append(usersToBan, strings.TrimRight(strings.TrimLeft(part, "<@!"), ">"))
-			continue
-		}
-		reasonMessage = strings.Join(args[i:], " ")
-		break
-	}
-	if len(usersToBan) == 0 {
-		return NewErrorEmbed("Ban Error", "You must specify which user(s) to ban.")
-	}
-	if messagesDaysToDelete > 7 {
-		return NewErrorEmbed("Ban Error", "You may only delete up to and including 7 days worth of messages from these users.")
-	}
-
-	if reasonMessage == "" {
-		for i := range usersToBan {
-			err := botData.DiscordSession.GuildBanCreate(env.Guild.ID, usersToBan[i], messagesDaysToDelete)
-			if err != nil {
-				return NewErrorEmbed("Ban Error", "An error occurred banning <@"+usersToBan[i]+">. Please consider manually banning and report this issue to a developer.")
-			}
-		}
-		return NewGenericEmbed("Ban", "Successfully banned the selected user(s).")
-	}
-	for i := range usersToBan {
-		err := botData.DiscordSession.GuildBanCreateWithReason(env.Guild.ID, usersToBan[i], reasonMessage, messagesDaysToDelete)
-		if err != nil {
-			return NewErrorEmbed("Ban Error", "An error occurred banning <@"+usersToBan[i]+">. Please consider manually banning and report this issue to a developer.")
-		}
-	}
-	return NewGenericEmbed("Ban", "Successfully banned the selected user(s) for the following reason:\n**"+reasonMessage+"**")
-}
-
-func commandSettingsServer(args []string, env *CommandEnvironment) *discordgo.MessageEmbed {
-	switch args[0] {
-	case "joinmsg":
-		guildSettings[env.Guild.ID].UserJoinMessage = strings.Join(args[1:], " ")
-		guildSettings[env.Guild.ID].UserJoinMessageChannel = env.Channel.ID
-		return NewGenericEmbed("Server Settings - Join Message", "Successfully set the join message to this channel.")
-	case "leavemsg":
-		guildSettings[env.Guild.ID].UserLeaveMessage = strings.Join(args[1:], " ")
-		guildSettings[env.Guild.ID].UserLeaveMessageChannel = env.Channel.ID
-		return NewGenericEmbed("Server Settings - Leave Message", "Successfully set the leave message to this channel.")
-	case "reset":
-		switch args[1] {
-		case "joinmsg":
-			guildSettings[env.Guild.ID].UserJoinMessage = ""
-			guildSettings[env.Guild.ID].UserJoinMessageChannel = ""
-		case "leavemsg":
-			guildSettings[env.Guild.ID].UserLeaveMessage = ""
-			guildSettings[env.Guild.ID].UserLeaveMessageChannel = ""
-		}
-	}
-	return NewErrorEmbed("Server Settings Error", "Error finding the setting "+args[0]+".")
-}
-
-//Debug commands
-func commandBotInfo(args []string, env *CommandEnvironment) *discordgo.MessageEmbed {
-	botGuilds, err := botData.DiscordSession.UserGuilds(100, "", "")
-	if err != nil {
-		return NewErrorEmbed("Bot Info Error", "An error occurred retrieving a list of guilds.")
-	}
-	botGuildNames := make([]string, 0)
-	for i := 0; i < len(botGuilds); i++ {
-		botGuildNames = append(botGuildNames, botGuilds[i].Name)
-	}
-
-	return NewEmbed().
-		SetTitle("Bot Info").
-		SetDescription("Info regarding this bot.").
-		AddField("Guild List ("+strconv.Itoa(len(botGuildNames))+")", strings.Join(botGuildNames, "\n")).
-		SetColor(0x1C1C1C).MessageEmbed
 }
